Header/Footer fixes to facilitate new FAQ pages
Altered all header/footer info to be a shared reference, so new pages can be added to the navigation bar without needing manual changes to every page.  Split up the getting-started page to flesh out each section more (and to not be such an info dump).  Added guides on buying/selling cards.